pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orange3-nlp"
version = "0.0.3"
description = "A collection of Orange3 widgets to perform natural language processing"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "CC-BY-NC-SA-4.0" }
authors = [
    { name = "Chris Lee", email = "github@chrislee.dhs.org" },
]
keywords = ["orange3 add-on"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3 :: Only"
]
dependencies = [
    "Orange3 >=3.34.0",
    "orange-canvas-core >=0.1.28",
    "orange-widget-base >=4.20.0",
    "spacy >= 3.8.5",
    "flair >= 0.15.1",
    "nltk >= 3.9.1",
    "numpy == 1.26.4",
    "sumy >= 0.11.0",
    "summa >= 1.2.0",
    "stanza >= 1.10.1",
    "ufal.udpipe >= 1.3.1.1",
    "faiss-cpu >= 1.11.0",
    "sentence_transformers >= 4.1.0",
    "openai >= 1.78.1",
    "langchain-text-splitters >= 0.3.8",
    "gensim >= 4.3.3",
    "tensorflow >= 2.19.0",
    "tensorflow-hub >= 0.16.1",
    "fasttext >= 0.9.3"
]

[license]
license = "CC-BY-NC-SA-4.0"
license-files = ["LICENSE"]

[project.optional-dependencies]
test = ["coverage"]
doc = ["sphinx", "recommonmark", "sphinx_rtd_theme"]

[project.urls]
Homepage = "https://github.com/chrislee35/orange3-nlp"

[tool.setuptools.packages.find]
where = ["."]
namespaces = true

[tool.setuptools.package-data]
"orangecontrib.nlp.widgets" = ["icons/*.svg"]

[tool.setuptools.data-files]
"help/orange3-nlp" = ["doc/_build/html/**/*"]

[project.entry-points."orange.widgets"]
nlp = "orangecontrib.nlp.widgets"
embedders = "orangeconfig.nlp.widgets.embedders"

[project.entry-points."orange3.addon"]
orange3-nlp = "orangecontrib.nlp"

[project.entry-points."orange.canvas.help"]
html-index = "orangecontrib.nlp.widgets:WIDGET_HELP_PATH"
